requirements.txt
aiogram
aiohttp
async-timeout
attrs
certifi
cffi
python-dotenv
pyTelegramBotAPI
python-telegram-bot